Refactor dataview & download with single-spa changes also ensure we forceUpdate app on RequestPluginRerenderType
