+++ packages/datagateway-download/src/App.tsx
@@ -9,7 +9,11 @@ import {
   StylesProvider,
 } from '@material-ui/core/styles';
 import ConfigProvider from './ConfigProvider';
-import { Preloader } from 'datagateway-common';
+import {
+  MicroFrontendId,
+  Preloader,
+  RequestPluginRerenderType,
+} from 'datagateway-common';
 import { DGThemeProvider } from 'datagateway-common';
 import AdminDownloadStatusTable from './downloadStatus/adminDownloadStatusTable.component';
 
@@ -26,6 +30,30 @@ class App extends Component<unknown, { hasError: boolean }> {
   public constructor(props: unknown) {
     super(props);
     this.state = { hasError: false };
+
+    this.handler = this.handler.bind(this);
+  }
+
+  handler(e: Event): void {
+    // attempt to re-render the plugin if the corresponding div is present
+    const action = (e as CustomEvent).detail;
+    if (action.type === RequestPluginRerenderType) {
+      // This is a temporary fix for the current issue with the tab indicator
+      // not updating after the size of the page has been altered.
+      // This is issue is being tracked by material-ui (https://github.com/mui-org/material-ui/issues/9337).
+      setTimeout(() => {
+        window.dispatchEvent(new CustomEvent('resize'));
+      }, 125);
+      this.forceUpdate();
+    }
+  }
+
+  public componentDidMount(): void {
+    document.addEventListener(MicroFrontendId, this.handler);
+  }
+
+  public componentWillUnmount(): void {
+    document.removeEventListener(MicroFrontendId, this.handler);
   }
 
   public componentDidCatch(error: Error | null): void {
